Adds UserRepository interface
