+++ bot/src/main/java/edu/java/bot/repository/UserRepository.java
@@ -1,22 +1,11 @@
 package edu.java.bot.repository;
 
 import edu.java.bot.entity.User;
-import java.util.HashMap;
 import java.util.Optional;
 
-public class UserRepository {
+public interface UserRepository {
 
-    private final HashMap<Long, User> users;
+    Optional<User> findUserById(long id);
 
-    public UserRepository() {
-        users = new HashMap<>();
-    }
-
-    public Optional<User> findUserById(long id) {
-        return users.containsKey(id) ? Optional.of(users.get(id)) : Optional.empty();
-    }
-
-    public void saveUser(User user) {
-        users.put(user.getId(), user);
-    }
+    void saveUser(User user);
 }
